pyproject.toml
[project]
name = "photoexpress"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "aiogram>=3.20.0.post0",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.41",
]
